update demo project
