+++ sdk/monitor/Azure.Monitor.OpenTelemetry.Exporter/tests/Azure.Monitor.OpenTelemetry.Exporter.Demo/Azure.Monitor.OpenTelemetry.Exporter.Demo.csproj
@@ -6,6 +6,16 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <None Remove="OTEL_DIAGNOSTICS.json" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Content Include="OTEL_DIAGNOSTICS.json">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </Content>
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="Azure.Identity" VersionOverride="1.8.0" />
     <PackageReference Include="Microsoft.Extensions.Logging" VersionOverride="6.0.0" />
     <PackageReference Include="OpenTelemetry.Extensions.AzureMonitor" />
