+++ sdk/monitor/Azure.Monitor.OpenTelemetry.Exporter/tests/Azure.Monitor.OpenTelemetry.Exporter.Demo/Logs/LogDemo.cs
@@ -1,8 +1,6 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#nullable disable // TODO: remove and fix errors
-
 using System;
 using Azure.Core;
 using Microsoft.Extensions.Logging;
@@ -13,7 +11,7 @@ namespace Azure.Monitor.OpenTelemetry.Exporter.Demo.Logs
     {
         private readonly ILoggerFactory loggerFactory;
 
-        public LogDemo(string connectionString, TokenCredential credential = null)
+        public LogDemo(string connectionString, TokenCredential? credential = null)
         {
             this.loggerFactory = LoggerFactory.Create(builder =>
             {
